backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
pandas
langchain-groq
langchain-core
pydantic
python-docx
python-pptx
requests
google-generativeai
python-multipart